+++ CHANGELOG.markdown
@@ -1,3 +1,8 @@
+# 0.0.3.pre
+
+* Removed dynamic ActiveRecord caller instrumentation
+* Fixed issue that prevents the app from loading if ActiveRecord isn't used.
+
 # 0.0.2
 
 Doesn't prevent app from loading if no configuration exists for the current environment.